Dados de emissão de CO2 do solo e atributos relacionados

Base de dados contendo valores de emissão de CO2 e atributos físicos e químicos do solo para diferentes experimentos conduzidos no Brasil

data_fco2

Format

Uma tibble, contendo 15397 linhas e 39 colunas

experimento

A classificação Espacial refere-se a experimentos nos quais uma malha amostral foi instalada como o objetivo de avaliar a variabilidade espacial da emissão de \(CO_2\) e atributos do solo relacionados por meio de técnicas geoestatísticas. Nos ensaios classificados com Temporal, a respiração do solo foi avaliada em diferentes tratamentos (usos, manejos ou preparo do solo).

data

Data da realização da avaliação em questão.

manejo

Manejo ou preparo do solo ou uso da terra que a parcela experimental foi submetida.

tratamento

Especificação a respeito do manejo/preparo/uso: AD (arado de disco); E (escarificador); GN ou gn (grade niveladora); GA (grade aradora); CQ (cana queimada); CC (cana crua); PD (plantio direto); EU (eucalipto); MT (mata ciliar); MN (mata natural); PI (pinus); SI (silvipastoril); PDr (pastagem degradada) e PR (pastagem reformada).

revolvimento_solo

Variável lógica indicando se houve revolvimento do solo previamente à condução do ensaio em questão.

data_preparo

Data em que ocorreu o revolvimento do solo, se verdadeiro.

conversao

Data em que a área foi convertidada de vegetação natural para o uso agrícola.

cobertura

Variável lógica indicando se no momento de availação o solo encontrava-se com cobertura superficial.

cultura

Cultura agrícola, florestal ou natural a que se destina a área.

x

coordenada X relativa à parcela ou malha amostral específica.

y

coordenada Y relativa à parcela ou malha amostral específica.

longitude_muni

Coordenada referente ao município no qual o experimento foi conduzido.

latitude_muni

Coordenada referente ao município no qual o experimento foi conduzido.

estado

Unidade federativa do município.

municipio

Cidade/município onde o experiemnto foi realizado.

ID

Identificação do ponto amostral dentro do ensaio específico.

prof

profunidade de coleta do atributo.

FCO2

Emissão de \(CO_2\) do solo em \(\mu mol\;m^{-2}\; s^{-1}\)

Ts

Temperatura do solo em \(^{o}C\).

Us

Umidade do solo em \(\%\).

pH

Potencial hidrogeniônico da solução do solo.

MO

Teor de matéria orgânica do solo em \(g \; dm^{-3}\).

P

Teor de Fósforo disponível do solo em \(mg \; dm^{-3}\).

K

Teor de Potássio do solo em \(mmol_c \; dm^{-3}\).

Ca

Teor de Cálcio do solo em \(mmol_c \; dm^{-3}\).

Mg

Teor de Magnésio do solo em \(mmol_c \; dm^{-3}\).

H_Al

Teor de H+Al do solo em \(mmol_c \; dm^{-3}\).

SB

Soma de bases do solo em \(mmol_c \; dm^{-3}\).

CTC

Capacidade de troca caiônica do soloem \(mmol_c \; dm^{-3}\).

V

Saturação por bases do solo em \(\%\).

Ds

Densidade do solo em \(g \; cm^{-3}\).

Macro

Macroporosidade do solo em \(m^{3}m^{-3}\).

Micro

Microporosidade do solo em \(m^{3}m^{-3}\).

VTP

Volume total de poros do solo em \(m^{3}m^{-3}\).

PLA

Porosidade livre de água do solo em \(\%\).

AT

Teor de areia total do solo em \(g \; kg^{-1}\).

SILTE

Teor de silte do solo em \(g \; kg^{-1}\).

ARG

Teor de argila do solo em \(g \; kg^{-1}\).

HLIFS

Grau de humificação da matéria orgânica do solo em unidade arbitrária

Source

www.fcav.unesp.br